format: divide every detailed components
