+++ src/App.tsx
@@ -1,14 +1,16 @@
 import React, { useEffect, useState } from 'react';
-
 import { Box } from '@chakra-ui/react';
 
 import DataTable from './components/DataTable';
-import ThemeToggleButton from './components/partials/ThemeToggleButton';
-import SearchWrapper from './components/SearchWrapper';
+import ThemeToggleButton from './components/globalpartials/ThemeToggleButton';
 import BodyWrapper from './components/BodyWrapper';
 import PageWrapper from './components/PageWrapper';
 import { SubareaInterface } from './const/types';
 import { getAllSubarea } from './Data/Api';
+import FunctionalWrapper from './components/FunctionalWrapper';
+import TableTitle from './components/functionalcomponents/TableTitle';
+import GlobalSearch from './components/functionalcomponents/GlobalSearch';
+import SearchButton from './components/functionalcomponents/SearchButton';
 
 function App(): JSX.Element {
   const [tableData, setTableData] = useState<SubareaInterface>();
@@ -25,7 +27,12 @@ function App(): JSX.Element {
     <Box>
       <DataTable>
         <Box>
-          <SearchWrapper />
+          <FunctionalWrapper>
+            <TableTitle>Member</TableTitle>
+            <GlobalSearch>
+              <SearchButton>Search</SearchButton>
+            </GlobalSearch>
+          </FunctionalWrapper>
           <BodyWrapper tabledata={tableData} />
           <PageWrapper />
         </Box>
