+++ src/components/functionalcomponents/InputContext.tsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const InputContext = createContext<string>('');
