--- src/components/partials/Rating.tsx
@@ -1,33 +1,0 @@
-import React from 'react';
-import { FaStar } from 'react-icons/fa';
-
-import { HStack, Icon, StackProps, useColorModeValue } from '@chakra-ui/react';
-
-interface Props {
-  defaultValue?: number;
-  max?: number;
-  size?: 'sm' | 'md' | 'lg' | 'xl';
-  rootProps?: StackProps;
-}
-
-function Rating(props: Props) {
-  const { defaultValue = 0, max = 5, size = 'md', rootProps } = props;
-  const color = useColorModeValue('gray.200', 'gray.600');
-  const activeColor = useColorModeValue('purple.500', 'purple.200');
-  return (
-    <HStack spacing="0.5" {...rootProps}>
-      {Array.from({ length: max })
-        .map((_, index) => index + 1)
-        .map((index) => (
-          <Icon
-            key={index}
-            as={FaStar}
-            fontSize={size}
-            color={index <= defaultValue ? activeColor : color}
-          />
-        ))}
-    </HStack>
-  );
-}
-
-export default Rating;
